backend/requirements.txt
flask==2.3.0
boto3==1.28.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0

